hal: Determine ACDB ID base on audio usecase ID on automotive platform
Update audio HAL to use audio usecase ID instead of snd_device to determine
the ACDB device ID for DSP calibration on automotive platform. This is the
same mechanism used in audio HAL plugin for CODEC calibration. With this
update, for all the audio use cases, the ACDB device ID used for DSP and
CODEC calibration is consistent.

Change-Id: I7a315d8ac005df6da2f532ec34b9bda2d5599eb3
CRs-fixed: 834345